02/20/2024 update
Sprint 2 development and test for Keeping Track of Turn function. 

In CardGames.py, Development for keeping track of which player's turn it is in poker is added at the very bottom of the file below the Dealer class.  The turn_hist() function is to be added to a Poker Class.  Designed to that user input is to be placed in lieu of the statement "Player.name called/raised/folded."