Three of a Kind identification
Function for identifying a hand as three of a kind included in CardGames.py at the end of the file. 

Takes a hand as an argument, counts the occurrences of the ranks within the hand and compares them against the definition of three of a kind: 1 card of rank A, 1 card of rank B, and 3 cards of rank C. If the hand fits this definition, the name of the identification object is "Three of a Kind".  

Test and html report included in the test folder.