Created Show_Marshall_Hand Function
NOTE: All of first turn functionality has been commented out so that I could test out the show_marshall_hand functionality.

I was unable to run the entire game since multiple functions are in-progress and create a rabbithole of errors in our game. We should aim to patch things up and establish a working game after we merge all the functions together again.